% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_flag_or_scalar_integerish}
\alias{assert_flag_or_scalar_integerish}
\title{Assert Flag or Scalar Integerish}
\usage{
assert_flag_or_scalar_integerish(
  x,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{The object to check.}
}
\description{
Assertion function for \code{\link[=check_flag_or_scalar_integerish]{check_flag_or_scalar_integerish()}}. Will throw an error if
the input is invalid.
}
\keyword{internal}
