% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\docType{data}
\name{astring}
\alias{astring}
\title{Create pISA layer or *sk* component.}
\format{
An object of class \code{character} of length 1.
}
\usage{
astring
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projects', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (short layer name, Title,
Description, *ToDo: add fields*). See Examples.}

\item{class}{Assay class key string.
Possible values are 'EXP' and 'MODEL'.}

\item{file}{File name with path, relative to layer.}
}
\value{
FAIRDOMhub created component.
}
\description{
Create pISA layer or *sk* component.
}
\note{
Upon success (status code 200) details
of newly created component can be used. Check status code.

Created component id is set in options.
}
\seealso{
\code{\link{skGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
