% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{compute_newline_locs}
\alias{compute_newline_locs}
\title{Locate newline character positions in a string}
\usage{
compute_newline_locs(text)
}
\arguments{
\item{text}{Text content as a single string.}
}
\value{
Integer matrix of newlines position with an additional row for the first
line where the start and end position are 0.
}
\description{
Computes the ending positions of each line (handling all newline variants).
}
\keyword{internal}
