% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{contentType}
\alias{contentType}
\title{Determine MIME type for file.}
\usage{
contentType(file)
}
\arguments{
\item{file}{file name.}
}
\value{
MIME type string.
}
\description{
Determine MIME type for file.
}
\examples{
contentType("bla.txt")
contentType("bla.pdf")
contentType("bla.tar.gz")
contentType("bla.bla")
contentType("bla")
contentType("bla.")

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
