% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{expand}
\alias{expand}
\title{Expand shortened type name into full name.}
\usage{
expand(type)
}
\arguments{
\item{type}{Component name (possibly shortened e.g. 'pe' for 'people',
'proj' for 'projects', ...).}
}
\value{
Expanded type name.
}
\description{
Expand shortened type name into full name.
}
\examples{
expand("pe")
expand("proj")
expand("inv")
expand("st")
expand("a")
expand(c("a","inv"))
# expand("p")
# Error in expand("p") : No such type or duplicate entry for 'p'

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
