% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_lower_file_extension}
\alias{extract_lower_file_extension}
\title{Extract Lowercase File Extensions}
\usage{
extract_lower_file_extension(files)
}
\arguments{
\item{files}{A character vector of files to search (only for \code{seek_in()}).}
}
\value{
A character vector of lowercase file extensions.
}
\description{
Extracts the file extensions from the provided file paths, normalizes them
to lowercase, and returns them as a character vector. The extension includes
the leading period (\code{.}).
}
\keyword{internal}
