% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhData}
\alias{fhData}
\title{Get content from an *fh* object.}
\usage{
fhData(r, part, ...)
}
\arguments{
\item{r}{Object retrieved by fhGET.}

\item{part}{Name of the required element. If missing, a list with
all relevant objects is returned.}
}
\value{
File name (string).
}
\description{
Get content from an *fh* object.
}
\note{
Parameter ... is ignored at this time.
}
\examples{
\\donotrun{
fhIni()
options()$fhub$myid
r <- fhGET("people",options()$fhub$myid)
d <- fhData(r)
names(d)
d$last_name
fhData(r)$tools
# Get list of people
r <- fhGET("people")
d <- fhData(r)
length(d)
names(d)
names(d[[1]])
titles <- sapply(d,function(x) x$attributes$title)
head(titles)
# Get FAIRDOMhub user id
myname <- titles[1]
myname
d[[pmatch(myname,titles)]]
id <- d[[pmatch(myname,titles)]]$id
id
}
}
\seealso{
\code{\link{get ...}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
