% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhFindId}
\alias{fhFindId}
\title{Get details of component with id from an *fh* object.}
\usage{
fhFindId(part, title)
}
\arguments{
\item{part}{Components name (e.g. 'people', 'projets', ...).}

\item{title}{Character string with the identifier
of the component (Title part).}
}
\value{
FAIRDOMhub component identifier: id, type and title.
    If argument title is missing, a data frame with identifiers
    for all items is returned.
}
\description{
Get details of component with id from an *fh* object.
}
\examples{
\\donotrun{
fhIni()
id <- fhFindId("people","Guest")
id
fhFindTitle("people",id)
# List of projects
projects <- fhFindId("projects")
head(projects)
}
}
\seealso{
\code{\link{fhFindTitle}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
