% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhGET}
\alias{fhGET}
\title{Get inormation from repository.}
\usage{
fhGET(what, id, uri = options()$fhub$baseurl, ...)
}
\arguments{
\item{what}{Type of information (e.g. "person").}

\item{id}{Repository id of an item.}

\item{uri}{Repository base address (URI)..}
}
\value{
File name (string).
}
\description{
Get inormation from repository.
}
\note{
Parameter ... is ignored at this time.
}
\examples{
\\donotrun{
fhIni()
options()$fhub$myid
r <- fhGET("people",options()$fhub$myid)
names(r)
r$status_code
}
}
\seealso{
\code{\link{get ...}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
