% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhIni}
\alias{fhIni}
\title{Initialize FAIRDOMhub information}
\usage{
fhIni(test = TRUE)
}
\arguments{
\item{test}{If TRUE, test server will be used..}
}
\value{
A list with URL and user information. For side effect see Notes.
}
\description{
Define FAIRDOMhub URL and user data
}
\examples{
\dontrun{
fhIni()
options("fhub")
#
fhIni(FALSE)
options("fhub")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
