% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhParse}
\alias{fhParse}
\title{Parse the response from SEEK API}
\usage{
fhParse(resp, ...)
}
\arguments{
\item{resp}{Response from SEEK API.}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
Parse the response from SEEK API
}
\examples{
\\donotrun{
fhIni()
options()$fhub$myid
r <- fhGet("people",options()$fhub$myid)
names(r)
r$response$status_code
status_code(r$response)
names(r$content)
r
r <- fhGet("people")
length(r$content)
names(r$content)
r$content[[1]]
r
}
}
\seealso{
\code{\link{fhGget ...}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
