% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhSkeleton}
\alias{fhSkeleton}
\title{Create *fh* skeleton.}
\usage{
fhSkeleton(type = "assay", meta)
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projets', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (Title, Description, *ToDo: add fields*).}
}
\value{
A list with the minimal information structure.
}
\description{
Creates *fh* object with required structure.
}
\seealso{
\code{\link{fhCreate}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
