% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{fileName}
\alias{fileName}
\title{Extract file name}
\usage{
fileName(x, ...)
}
\arguments{
\item{x}{Complete file path or  file.type name.}

\item{...}{Any other arguments.}
}
\value{
File name (string).
}
\description{
Extract file name from a file path.
}
\note{
Parameter ... is ignored at this time.
}
\examples{
fileName(".\\\\validation.Rnw")
fileName("./bla/validation.Rnw")
fileName("./validation.")
fileName("./validation")
}
\seealso{
\code{\link{fileType}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
