% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{fileType}
\alias{fileType}
\title{Extract file type}
\usage{
fileType(x, ...)
}
\arguments{
\item{x}{Complete file path or  file name.}

\item{...}{Any other arguments.}
}
\value{
File type (string).
}
\description{
Extract file type from a file path.
}
\note{
Parameter ... is ignored at this time.
}
\examples{
fileName(".\\\\validation.Rnw")
fileName("./bla/validation.Rnw")
fileName("./validation.")
fileName("./validation")
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
