% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{fsummary}
\alias{fsummary}
\title{As factor summary of a data frame}
\usage{
fsummary(x, ...)
}
\arguments{
\item{x}{Data frame.}

\item{...}{Any other arguments.}
}
\value{
Summary object.
}
\description{
As factor summary of a data frame
}
\note{
Argument ... not used
}
\examples{
fsummary(data.frame(x=rnorm(20),txt=sample(letters,20,rep=TRUE)))
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{summary}
