% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{getLayer}
\alias{getLayer}
\title{Get pISA layer name}
\usage{
getLayer(x, path = getwd())
}
\arguments{
\item{x}{Layer character (one of p, I, S, or A).}

\item{path}{Directory path, defaults to working directory.}
}
\value{
Character string with layer name.
}
\description{
Get pISA layer name
}
\examples{
\dontrun{
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/other"
oldwd <- setwd(system.file("extdata",astring,package="pisar"))
oldwd
.pname <- getLayer("p")
.pname
getLayer("I")
getLayer("S")
getLayer("A")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
