% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{getRoot}
\alias{getRoot}
\title{Get root directory for pISA layer}
\usage{
getRoot(x = "p", path = getwd(), ...)
}
\arguments{
\item{x}{character characteristic for pISA layer}

\item{path}{path within the pISA-tree}

\item{...}{any other arguments}
}
\value{
data frame with Key/value pairs
}
\description{
Get root directory for pISA layer
}
\note{
Argument ... not used.
}
\examples{
getRoot("p", path="d:/_p_prj/_I_inv/_S_st/_A_asy/other/doc")  
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pISA}
