% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{glue_escape}
\alias{glue_escape}
\title{Escape curly braces for use in glue::glue templates}
\usage{
glue_escape(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Escaped character vector
}
\description{
Converts \verb{\{} to \verb{\{\{} and \verb{\}} to \verb{\}\}} to prevent premature evaluation in \code{glue::glue()}.
}
\keyword{internal}
