% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-files.R
\name{is_in_gitfolder}
\alias{is_in_gitfolder}
\title{Check if Files Are Located in a \code{.git} Folder}
\usage{
is_in_gitfolder(files)
}
\arguments{
\item{files}{A character vector of files to search (only for \code{seek_in()}).}
}
\value{
A logical vector indicating whether each file is located within a \code{.git} folder.
}
\description{
Identifies whether the provided file paths are located inside a \code{.git} directory.

This function assumes that the file paths are normalized beforehand (i.e., using
forward slashes \code{/} even on Windows systems).
}
\keyword{internal}
