% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{out.path}
\alias{out.path}
\title{Create output directory}
\usage{
out.path(out.dir = "../out", args = "", which = 1:length(args))
}
\arguments{
\item{out.dir}{Character string, base output directory.}

\item{args}{Character vector, arguments used for sub-analysis.
or from a batch call}

\item{which}{Numeric vector, which arguments to use.}
}
\value{
Directory name.
}
\description{
Create output directory, name it with appended arguments.
}
\note{
Directory is created
}
\examples{
out.path()
out.path(args="")
dir.exists("../out/xx-ena-dva")
out.path(args=c("xx.txt","ena","dva"))
dir.exists("../out/xx-ena-dva")
unlink("../out/xx-ena-dva",recursive=TRUE)
dir.exists("../out/xx-ena-dva")
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
