% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{parse_files_to_dfs}
\alias{parse_files_to_dfs}
\title{Read Files and Return Content as Data Frames}
\usage{
parse_files_to_dfs(files, warn, n)
}
\arguments{
\item{files}{A character vector of file paths to read.}

\item{warn}{logical.  Warn if a text file is missing a final \abbr{EOL} or if
    there are embedded \abbr{NUL}s in the file.}

\item{n}{integer.  The (maximal) number of lines to
    read.  Negative values indicate that one should read up to the end of
    input on the connection.}
}
\value{
A list of data frames, each representing the content of a file. Each data frame includes columns for file index, path, line number, and content.
}
\description{
Reads a list of files and returns a list of data frames, one per file, each containing
the content of the file by line.
}
\keyword{internal}
