% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{pisa}
\alias{pisa}
\title{Extract pISA-tree details}
\usage{
pisa(path = getwd(), addArgs = NULL, global = TRUE)
}
\arguments{
\item{path, }{defaults to working directory}

\item{addArgs}{character vector, additional arguments}

\item{global}{if \code{TRUE} (default) auxiliary objects will be
created in the global environment (see note)}
}
\value{
a list with layer information components, possibly changing 
objects in the global environment (Se Note).
}
\description{
Extract pISA-tree details: name, root and metadata
    for all layers above the current directory (below or in the Assay).
}
\note{
If argument global is \code{TRUE} (default), auxiliary objects with
    pISA related information will be created in the global environment.
    Sometimes it is more convenient to use such object instead 
    of the elements of the (invisibly) returned pisa list. 
    The object are hidden (names start with dot); use
    \code{ls(pattern="^\\.",all.names=TRUE)}
    to get a full list of hiddent objects. The created objects are
\describe{
    \item{.[pisa]name}{layer name}
    \item{.[pisa]root}{layer path (relative to the working directory)}
    \item{.[pisa]meta}{data frame with corresponding layer metadata}
    \item{.oroot}{output directory path}
    \item{.inroot}{input (data) directory path}
    \item{.reproot}{report directory path}
    \item{.reproot}{report directory path}
    \item{.outputFile}{output file path and name}
    \item{.args}{a vector of additional arguments (possibly from a batch call)}
    \item{.pfn}{phenodata file path and name (relative to Investigation)}
    \item{.ffn}{featuredata fiel path and name}
    \item{.outfn}{output file basename (no type)}
    \item{.rnwfn}{knitr source file name (*.Rnw or *.Rmd)}
}
}
\examples{
\dontrun{
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/other"
oldwd <- setwd(system.file("extdata",astring,package="pisar"))
oldwd
pisa <- pisa(global=FALSE)
str(pisa)
names(pisa)
dir(pisa$p$root)
setwd(oldwd)
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
