% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{print.pISAmeta}
\alias{print.pISAmeta}
\title{Print metadata object as Dlist}
\usage{
\method{print}{pISAmeta}(x, width = max(nchar(x[, 1])) * 3.5, ...)
}
\arguments{
\item{x}{Metadata object, data.frame with two columns.}

\item{width}{Estimated text width.}

\item{...}{Any other arguments.}
}
\description{
Print metadata object as Dlist
}
\note{
Metadata table is printed in convenient Dlist form.
}
\examples{
.pISAloc <- system.file("extdata","_p_Demo",package="pisar")
readMeta(.pISAloc)
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{package}
