% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{field}
\alias{field<-}
\alias{fields}
\alias{regex}
\alias{fixed}
\alias{coll}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stringr}{\code{\link[stringr:coll]{coll()}}, \code{\link[stringr:fixed]{fixed()}}, \code{\link[stringr:regex]{regex()}}}

  \item{vctrs}{\code{\link[vctrs:field]{field()}}, \code{\link[vctrs:field<-]{field<-()}}, \code{\link[vctrs:fields]{fields()}}}
}}

