% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-package.R
\docType{package}
\name{seekr-package}
\alias{seekr-package}
\title{seekr: Search, Inspect, and Replace Text Across Files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An inspectable and composable workflow for search-and-replace in text files. Files can be listed, filtered, and searched separately, with inspectable exclusions showing what was excluded and why. Matches are represented as structured vectors that can be printed with context, summarized, and filtered. Replacements can be defined at search time or set and updated after the search. Only matches present in the vector are modified when files are written. Backup and restore helpers are provided for file workflows. Text that has already been read can also be searched and updated directly, giving users control over input, output, and encoding when needed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/smartiing/seekr}
  \item \url{https://smartiing.github.io/seekr/}
  \item Report bugs at \url{https://github.com/smartiing/seekr/issues}
}

}
\author{
\strong{Maintainer}: Sacha Martingay \email{martingay.sacha@hotmail.com} [copyright holder]

Authors:
\itemize{
  \item Sacha Martingay \email{martingay.sacha@hotmail.com} [copyright holder]
}

}
\keyword{internal}
