% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{setPublicAccess}
\alias{setPublicAccess}
\title{Set public access permission.}
\usage{
setPublicAccess(meta, na.access = "no_access")
}
\arguments{
\item{meta}{data frame or list. pISA metadata object.}

\item{na.access}{character string. Default access, either 'no_access' for #'      Private layer or 'download' for Public layer.}
}
\description{
Set public access permission according to the "Sharing permission:"
key in metadata.
}
\examples{
meta = data.frame(
Key=c("Assay:","Sharing permission:"), Value=c("_A_Test","Public"))
setPublicAccess(meta)
meta = data.frame(
Key=c("Assay:","Sharing permission:"), Value=c("_A_Test","Private"))
setPublicAccess(meta)
meta = data.frame(
Key=c("Assay:"), Value=c("_A_Test"))
setPublicAccess(meta)
setPublicAccess(meta, "download")

}
