% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skCheck}
\alias{skCheck}
\title{Check if component exists within the parent layer.}
\usage{
skCheck(type, title, ptype, pid = skGetOption(ptype), verbose = FALSE)
}
\arguments{
\item{type}{Component type (possibly shortened e.g. 'pe' for 'people',
'proj' for 'projects', ...).}

\item{title}{Character string,  component name.}

\item{ptype}{Parent component type (possibly shortened).}

\item{pid}{Parent component id. If missing, appropriate part
skOptions is used.}

\item{verbose}{logical, if TRUE, details are printed.}
}
\value{
logical value indicating whether the component named by
     argument title exists.
}
\description{
This is a faster version of skExists, intended mostly to check existence 
    of ISA layers in the parent layer.
}
\note{
Check if named layer (.e.g. assay) exists in the parent layer.
    Parent layer is usually just one above (e.g. study where the assay
    is supposed to reside).
}
\examples{
\dontrun{
skReset()
skFindId("proj","_p_Demo")
type <- "inv"
title <- "_I_Test"
ptype <- "proj"
#system.time(print(skExists(type,title,ptype))) # 35 seconds
print.simple.list(skOptions("id"))
system.time(print(skCheck(type,title,ptype)))
print.simple.list(skOptions("id"))
type <- "inv"
title <- "Testa"
ptype <- "proj"
(skCheck(type,title,ptype))
       print.simple.list(skOptions("id"))
(skCheck("inv","Test layer","proj"))
       print.simple.list(skOptions("id"))
title <- "Test investigation"
(skCheck(type,title,ptype))
       print.simple.list(skOptions("id"))
type <- "inv"
ptype <- "proj"
title <- "_I_01_LabTrials-test"
       print.simple.list(skOptions("id"))
system.time(print(skCheck(type,title,ptype)))
system.time(skExists(type,title,ptype))
       print.simple.list(skOptions("id"))
type <- "stud"
title <- "_S_03_dsRNAprod"
ptype <- "inv"
(skCheck(type,title,ptype))
       print.simple.list(skOptions("id"))
#
type <- "assay"
title <- "_A_01_pIsol-wet"
ptype <- "stu"
(skCheck(type,title,ptype))
       print.simple.list(skOptions("id"))
skCheck("proj","_p_Demo","people")
}

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
