% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skContent}
\alias{skContent}
\title{Get content from an *sk* object.}
\usage{
skContent(r, node, ...)
}
\arguments{
\item{r}{object retrieved by skGet.}

\item{node}{name of the required element. If missing, a list with
all relevant objects is returned.}

\item{...}{further arguments.}
}
\value{
File name (string).
}
\description{
Get content from an *sk* object.
}
\note{
Parameter ... is ignored at this time.
}
\examples{
\dontrun{
options(.sk$test)
options("sk.myid")
r <- skRead("people",options("sk.myid"))
d <- skContent(r,"attributes")
names(d)
d$last_name
skContent(r)$attributes$tools
# Get list of people
r <- skList("people")
d <- skContent(r)
length(d)
names(d)
names(d[[1]])
titles <- sapply(d,function(x) x$attributes$title)
head(titles)
# Get FAIRDOMhub user id
myname <- titles[1]
myname
d[[pmatch(myname,titles)]]
id <- d[[pmatch(myname,titles)]]$id
id
}
}
\seealso{
\code{\link{get}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
