% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skCreate}
\alias{skCreate}
\title{Create pISA layer or *sk* component.}
\usage{
skCreate(type = "assays", meta = list(), class = "EXP", file = "NA.TXT")
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projets', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (Title,
Description, *ToDo: add fields*).}

\item{class}{Assay class key string.
Possible values are 'EXP' and 'MODEL'.}

\item{file}{File name with path, relative to layer.}
}
\value{
FAIRDOMhub created component.
}
\description{
Create pISA layer or *sk* component.
}
\note{
Upon success (status code 200) details
of newly created component can be used. Check status code.
}
\examples{
\dontrun{
if(FALSE)
{
options(.sk$test)
#skIni(prid = 26, test=TRUE)
options(sk.prid = 26)
skOptions("id")
 sp <- skCreate( type = "projects"
    , meta= list(
      Title=paste("Test project", Sys.time())
    , Description="Testing of upload")
    )
 str(sp)
 str(skData(sp))
# Add member manually
 #skIni(prid = 26, pid=104, test=TRUE)
 options(sk.prid=26, sk.pid=104)
 skOptions("id")
 si <- skCreate( type = "investigations"
  , meta= list(
      Title=paste("Test investigation", Sys.time())
    , Description="Testing of upload")
    )
 si
 skData(si)$id

 iid=skData(si)$id
 #skIni(prid = 26, pid=104, iid=iid, test=TRUE)
 options(sk.prid=26, sk.pid=104, sk.iid=iid)
 skOptions("id")
 ss <- skCreate( type = "studies"
  , meta= list(
      Title=paste("Test study", Sys.time())
    , Description="Testing of upload")
    )
 ss
 skData(ss)$id

 options(sk.prid = 26
       , sk.pid=104
       , sk.iid=skData(si)$id
       , sk.sid=skData(ss)$id
       )
 skOptions()
 sa <- skCreate( type = "assays"
  , meta= list(
      Title=paste("Test assay", Sys.time())
    , Description="Testing of upload")
    , class="EXP"
    )
 #str(sa)
 sa
}

# Type: data_file
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/"
oldwd <- setwd(system.file("extdata",astring,package="seekr"))
oldwd
.aname <- getLayer("A")
.aroot <- getRoot("A")
.ameta  <- readMeta()
 file <- "input/README.MD"
 options(sk.prid = 26, sk.pid=104, sk.iid=115 , sk.sid=117 , sk.aid=401)
 type <- "data_files"
 type <- "documents"
 sdat <- skCreate( type = type
  , meta= list(
      Title=paste("Test assay", Sys.time())
    , Description="Testing of upload")
  , file=file
    )
 #str(sdat)
 sdat
 skData(sdat)$id
if(interactive()) setwd(oldwd)
getwd()
res <- sdat$content
item_link <- file.path(res$meta$base_url,res$links$self)
if(interactive()) shell.exec(item_link)
}
}
\seealso{
\code{\link{skGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
