% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skData}
\alias{skData}
\title{Get content from an *sk* object.}
\usage{
skData(r, node, ...)
}
\arguments{
\item{r}{Object retrieved by skGet.}

\item{type}{Name of the required element. If missing, a list with
all relevant objects is returned.}
}
\value{
File name (string).
}
\description{
Get content from an *sk* object.
}
\note{
Parameter ... is ignored at this time.
}
\examples{
\donotrun{
skIni()
options()$fhub$myid
r <- skGet("people",options()$fhub$myid)
d <- skData(r,"attributes")
names(d)
d$last_name
skData(r)$tools
# Get list of people
r <- skGet("people")
d <- skData(r)
length(d)
names(d)
names(d[[1]])
titles <- sapply(d,function(x) x$attributes$title)
head(titles)
# Get FAIRDOMhub user id
myname <- titles[1]
myname
d[[pmatch(myname,titles)]]
id <- d[[pmatch(myname,titles)]]$id
id
}
}
\seealso{
\code{\link{get ...}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
