% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skDelete}
\alias{skDelete}
\title{Delete component.}
\usage{
skDelete(type, id, uri = options("sk.url"), ...)
}
\arguments{
\item{type}{Component type (e.g. "assay").}

\item{id}{Repository id of component.}

\item{uri}{Repository base address (URI).}

\item{...}{further arguments.}
}
\value{
Deleted object (list) of class \code{seek_api}.
}
\description{
Delete component.
}
\note{
Parameter ... is ignored at this time.

Component id in options is set to NULL.
}
\examples{
\dontrun{
options(.sk$test)
options("sk.myid")
r <- skDelete("assay",options("sk.aid"))
names(r)
r$response$status_code
status_code(r$response)
r
}
}
\seealso{
\code{\link{skCreate}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
