% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skExists}
\alias{skExists}
\title{Check if component exists within the layer.}
\usage{
skExists(type, title, ptype, pid, verbose = FALSE)
}
\arguments{
\item{type}{Component type (possibly shortened e.g. 'pe' for 'people',
'proj' for 'projects', ...).}

\item{title}{Character string,  component name.}

\item{ptype}{Parent component type (possibly shortened).}

\item{pid}{Parent component id.}

\item{verbose}{logical, if TRUE, details are printed.}
}
\value{
logical value indicating whether the component named by
     argument title exists.
}
\description{
Check if component exists within the layer.
}
\note{
Check if named layer (.e.g. assay) exists in the parent layer.
    Parent layer is usually just one above (e.g. study where the assay 
    is supposed to reside) but can be any above.
}
\examples{
\dontrun{
options(.sk$test)
skFindId("projects","_p_Demo")
skOptions()
skExists("people", "Andrej Blejec", "proj", options("sk.pid"))
skExists("investigations","_I_Test", "proj", options("sk.pid"),TRUE)
skExists("pe", "Guest", "proj", options("sk.pid"))
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
