% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skExistsa}
\alias{skExistsa}
\title{Check if component exist anywhere in the repository.}
\usage{
skExistsa(type, title, verbose = FALSE)
}
\arguments{
\item{type}{Component type (possibly shortened e.g. 'pe' for 'people',
'proj' for 'projects', ...).}

\item{title}{Character string with the identifier
of the component (title part).}

\item{verbose}{logical, if TRUE, details are printed.}
}
\value{
logical value indicating whether the component named by
     argument title exists.
}
\description{
Check if component exist anywhere in the repository.
}
\examples{
\dontrun{
options(.sk$test)
skExistsa("people", "Andrej Blejec")
skExistsa("projects","Demo", verbose=TRUE)
skExistsa("pe", "Guest")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
