% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skExpand}
\alias{skExpand}
\title{Expand shortened type name into full name.}
\usage{
skExpand(type)
}
\arguments{
\item{type}{Component type (possibly shortened e.g. 'pe' for 'people',
'proj' for 'projects', ...).}
}
\value{
Expanded type name.
}
\description{
Expand shortened type name into full name.
}
\examples{
skExpand("pe")
skExpand("proj")
skExpand("inv")
skExpand("st")
skExpand("a")
skExpand(c("a","inv"))
# skExpand("p")
# Error in skExpand("p") : No such type or duplicate entry for 'p'

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
