% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skFindId}
\alias{skFindId}
\title{Get details of component with id from an *fh* object.}
\usage{
skFindId(type, title)
}
\arguments{
\item{type}{Components name (e.g. 'people', 'projets', ...).}

\item{title}{Character string with the identifier
of the component (title part).}
}
\value{
FAIRDOMhub component identifier: id, type and title.
    If argument title is missing, a data frame with identifiers
    for all items is returned.
}
\description{
Get details of component with id from an *fh* object.
}
\note{
If item is not found, value 0 is returned as id.
}
\examples{
\\donotrun{
skIni()
id <- skFindId("people","Guest")
id
skFindTitle("people",id)
# does not exist
skFindId("people","No User")
# List of projects
projects <- skFindId("projects")
head(projects)
}
}
\seealso{
\code{\link{skFindTitle}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
