% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skFindTitle}
\alias{skFindTitle}
\title{Get details of component with id from an *sk* object.}
\usage{
skFindTitle(type, id)
}
\arguments{
\item{type}{Components name (e.g. 'people', 'projects', ...).}

\item{id}{Character string with the identifier
of the component (id part).}
}
\value{
FAIRDOMhub component identifier: id, type and title.
    If argument title is missing, a data frame with identifiers
    for all items is returned. See note.
}
\description{
Get details of component with id from an *sk* object.
}
\note{
If item is not found, empty string is returned as title.

Touched component id is set in options.
}
\examples{
\dontrun{
skFindTitle("people",0)
}
}
\seealso{
\code{\link{skFindTitle}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
