% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skGet}
\alias{skGet}
\title{Get information from repository.}
\usage{
skGet(type, id, uri = options("sk.url"), verbose = FALSE, ...)
}
\arguments{
\item{type}{Type of information (e.g. "person").}

\item{id}{Repository id of an item.}

\item{uri}{Repository base address (URI).}

\item{verbose}{logical, print details if TRUE.}

\item{...}{further arguments.}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
Get information from repository.
}
\note{
Parameter ... is ignored at this time.

Touched component id is set in the options (see:
     \code{\link{skOptions}}).
}
\examples{
\dontrun{
options(.sk$test)
options("sk.myid")
r <- skGet("people",options("sk.myid"))
names(r)
r$response$status_code
status_code(r$response)
r
# Non existent user
skGet("people",0)
}
}
\seealso{
\code{\link{get}}, \code{\link{skRead}}, \code{\link{skList}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
