% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skGetLayer}
\alias{skGetLayer}
\title{Get layer name from path.}
\usage{
skGetLayer(layer = "p", path = getwd(), verbose = FALSE)
}
\arguments{
\item{layer}{character. pISA-tree layer prefix (one of characters "p",
"I", "S" or "A").}

\item{path}{Directory path, defaults to working directory.}

\item{verbose}{logical. Should warnings be printed?}
}
\description{
Get layer name from path.
}
\note{
This function should be included in the package pisar.
}
\examples{
\dontrun{
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/other"
oldwd <- setwd(system.file("extdata",astring,package="seekr"))
oldwd
.pname <- getLayer("p")
.pname
getLayer("I")
getLayer("S")
getLayer("A")
if(interactive()) setwd(oldwd)
getwd()
}

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa,}
\keyword{seekr}
