% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skGetOption}
\alias{skGetOption}
\title{Get seekr option according to the type.}
\usage{
skGetOption(type)
}
\arguments{
\item{type}{Components name (e.g. 'people', 'projects', ...).}
}
\value{
Option specified by type.
}
\description{
Get seekr option according to the type.
}
\examples{
\dontrun{
skGetOption("myid")
skSetOption("myid", 368)
skOptions()
skSetOption("assay",1234)
skGetOption("assay")
# Delete/reset option sk.aid
skSetOption("assay",NULL)
skOptions()
Get options
skGetOption("myid")
skGetOption("inv")
}
}
\seealso{
\code{\link{skSetOption}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
