% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skList}
\alias{skList}
\title{List operations return a list of all objects of the specified type,
to which the authenticated user has access.}
\usage{
skList(type, uri = options("sk.url"), ...)
}
\arguments{
\item{type}{Type of information (e.g. "person").}

\item{uri}{Repository base address (URI).}

\item{...}{further arguments.}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
List operations return a list of all objects of the specified type,
to which the authenticated user has access.
}
\note{
Parameter ... is ignored at this time.

Touched component id is set in the options (see:
     \code\link{skOptions}).
}
\examples{
\dontrun{
options(.sk$test)
options("sk.myid")
r <- skList("people")
names(r)
r$response$status_code
status_code(r$response)
l <- skContent(r)
length(l)
# Short list of users
head(sort(sapply(l, function(x) x$attributes$title)),10)
}
}
\seealso{
\code{\link{skRead}}, \code{\link{skGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
