% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skListp}
\alias{skListp}
\title{List of all objects of the specified type within the project.}
\usage{
skListp(
  class = projects,
  type,
  pid = getOption("sk.pid"),
  uri = options("sk.url"),
  ...
)
}
\arguments{
\item{class}{Type of parent layer (e.g. "projects").}

\item{type}{Type of information (e.g. "person").}

\item{pid}{Id of a project giving the scope of the list.}

\item{uri}{Repository base address (URI).}

\item{...}{further arguments.}
}
\value{
A data.frame with ids and titles of related object.
}
\description{
List of all objects of the specified type within the project.
}
\note{
Parameter ... is ignored at this time.

Touched component id is set in the options (see:
     \code{\link{skOptions}}).
}
\examples{
\dontrun{
options(.sk$test)
options("sk.pid")
r <- skListp("proj","people")
r
skListp("inv","assays")
skListp("proj", "inv")
}
}
\seealso{
\code{\link{skList}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
