% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skOptions}
\alias{skOptions}
\title{Show all seekr related options}
\usage{
skOptions(end = "", start = "sk.", hide = "sk.pwd")
}
\arguments{
\item{end}{character string defining the end of the string. By default 
all names are valid (empty string).}

\item{start}{character string defining the start of the name, 
default is "sk.", which is typical for seekr options.}

\item{hide}{vector of names, that should not be listed. 
Default is "sk.pwd" which prevents password to be revealed.}
}
\value{
A list of seekr related options.
}
\description{
Lists selected seekr related options. Selection is based on the start
and end of the name. All seekr options start with prefix '.sk'. 
Argument \code{end} give the required sufix for the option to be listed.
Typical use would be to list identifiers, when \code{end = "id"}.
}
\examples{
\dontrun{
#skIni()
options(.sk$test)
skOptions()
skOptions("id")
skOptions(hide=NULL)
}
}
\seealso{
\code{\link{startsWith}, \link{endsWith}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
