% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skParse}
\alias{skParse}
\title{Parse the response from SEEK API and convert to a list.}
\usage{
skParse(resp, ...)
}
\arguments{
\item{resp}{response from SEEK API.}

\item{...}{further arguments.}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
Parse the response from SEEK API and convert to a list.
}
\examples{
\dontrun{
options(.sk$test)
options("sk.myid")
r <- skRead("people",options("sk.myid"))
names(r)
r$response$status_code
status_code(r$response)
names(r$content)
r
r <- skRead("people")
length(r$content)
names(r$content)
r$content[[1]]
names(r)
r
}
}
\seealso{
\code{\link{skGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
