% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skRelated}
\alias{skRelated}
\title{List of relationships within an object.}
\usage{
skRelated(type, id, select = NULL, uri = options("sk.url"), ...)
}
\arguments{
\item{type}{Object type (e.g. "project").}

\item{id}{Id of the object giving the scope of the list.
If missing, appropriate part of skOptions is used.}

\item{select}{Vector of character strings with object types of interest.}

\item{uri}{Repository base address (URI).}

\item{...}{further arguments.}
}
\value{
A data.frame with ids, types, roles and titles of related objects.
}
\description{
List of relationships within an object.
}
\note{
Parameter ... is ignored at this time.

Id of each last touched component is set in the options (see:
     \code{\link{skOptions}}).
}
\examples{
\dontrun{
skRelated("projects",152)
skRelated("investigations",151)
skRelated("investigations",152)
skRelated("investigations",152, "assays")
skRelated("proj",options("sk.pid"), "inv")
skRelated("inv",152, c("people","studies"))
skRelated("inv",152, c("pe","a"))
skRelated("studies",165)
skRelated("assays",494)
skRelated("people",368,"projects")
}
}
\seealso{
\code{\link{skListp}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
