% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skRelationships}
\alias{skRelationships}
\title{List of relationships within an object.}
\usage{
skRelationships(type, id, select = NULL, uri = options("sk.url"), ...)
}
\arguments{
\item{type}{Object type (e.g. "project").}

\item{id}{Id of the object giving the scope of the list.}

\item{select}{Vector of character strings with object types of interest.}

\item{uri}{Repository base address (URI).}

\item{...}{further arguments.}
}
\value{
A data.frame with ids, types, roles and titles of related objects.
}
\description{
List of relationships within an object.
}
\note{
Parameter ... is ignored at this time.

Id of each last touched component is set in the options (see:
     \code{\link{skOptions}}).
}
\examples{
\dontrun{
skRelationships("projects",152)
skRelationships("investigations",151)
skRelationships("investigations",152)
skRelationships("investigations",152, "assays")
skRelationships("inv",152, c("people","studies"))
skRelationships("inv",152, c("pe","a"))
skRelationships("studies",165)
skRelationships("assays",494)
skRelationships("people",368,"projects")
}
}
\seealso{
\code{\link{skListp}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
