% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skReset}
\alias{skReset}
\title{Reset seekr related options}
\usage{
skReset(all = FALSE)
}
\arguments{
\item{all}{logical. Reset all options (including user and server related).}
}
\description{
Reset seekr related options
}
\examples{
\dontrun{
options(sk.myid=368, sk.pid=166, sk.usr="Username")
skOptions()
skReset()
skOptions()
}
}
\seealso{

}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
