% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skSearch}
\alias{skSearch}
\title{Search operation returns a list of references to resources.}
\usage{
skSearch(type = "data_files", q, uri = options("sk.url"), ...)
}
\arguments{
\item{type}{Type of information (e.g. "person", default is "data_files").}

\item{q}{Text to search for, spaces acceptable.}

\item{uri}{Repository base address (URI).}

\item{...}{further arguments.}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
Search operation returns a list of references to resources.
}
\note{
Parameter ... is ignored at this time.

Touched component id is set in the options (see:
     \code\link{skOptions}).
}
\examples{
\dontrun{
options(.sk$test)
options("sk.myid")
r <- skSearch("people","Guest")
names(r)
r$response$status_code
status_code(r$response)
str(skContent(r))
# Non existent user
x <- skSearch("people","XY")
x$content
str(skContent(x))
(skContent(skSearch("institutions","systems biology")))
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
