% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skSetLayers}
\alias{skSetLayers}
\title{Set seek layer names based on pISA-tree file path.}
\usage{
skSetLayers(
  path,
  root = skGetOption("root"),
  studyForFiles = "Investigation files",
  assaySuffix = "-files",
  verbose = FALSE
)
}
\arguments{
\item{path}{Directory path, defaults to working directory.}

\item{verbose}{logical. Should warnings be printed?}

\item{layer}{character. pISA-tree layer prefix (one of characters "p",
"I", "S" or "A").}
}
\description{
Set seek layer names based on pISA-tree file path.
}
\note{
SEEK allows only files to be associated with assays.
   To upload files in investigation and study levels, a dummy
   study with assays for each study will be used to hold their files.
}
\examples{
\dontrun{
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/other"
oldwd <- setwd(system.file("extdata",astring,package="seekr"))
oldwd
require(pisar)
pisa <- pisa()
options(.sk$test)
seekignore <- readLines(file.path(.iroot, "seekignore.txt"))
seekignore
dirs <- skFilesToUpload(.iroot)
skSetLayers(dirs[1], root=.iroot)
skSetLayers(dirs[2], root=.iroot)
skSetLayers(dirs[length(dirs)], root=.iroot)
skSetLayers(dirs[10], root=.iroot)
sapply(dirs, function(x) skSetLayers(x))
if(interactive()) setwd(oldwd)
getwd()
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa,}
\keyword{seekr}
