% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skUpload}
\alias{skUpload}
\title{Upload file.}
\usage{
skUpload(object, file)
}
\arguments{
\item{object}{A previously created SEEK component.
Must be one of 'documents' or 'data_files'.}

\item{file}{File name with path, relative to layer.}
}
\value{
FAIRDOMhub created component.
}
\description{
Upload file to a crerated object of type 'documents' or 'data_files'.
}
\note{
Upon success (status code 200) details
of newly created component can be used. Check status code.
}
\examples{
file <- dir()[1]
file
fhd <- skCreate("documents",meta,file=file)
fhd
sd <- skUpload(fhd, file)

}
\seealso{
\code{\link{skCreate}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
