% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skUploadFile}
\alias{skUploadFile}
\title{Upload file.}
\usage{
skUploadFile(
  file,
  root = skGetOption("root"),
  verbose = FALSE,
  reportFile = NULL,
  access = "no_access"
)
}
\arguments{
\item{file}{character. File name and path (from the root directory).}

\item{root}{character. Root directory.}
}
\description{
Create an object of type 'documents' or 'data_files'
and upload file content as a blob.
}
\note{
Argument file is usually a result of a function skFilesToUpload
     that sets the option 'root' that was used.
}
\examples{
\dontrun{
skUploadFile(file)
}

}
\seealso{
\code{\link{skFilesToUpload}}, \code{\link{skUpload}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
