% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skUploadFiles}
\alias{skUploadFiles}
\title{Create appropriate layers and upload files.}
\usage{
skUploadFiles(
  files,
  root = skGetOption("root"),
  test = FALSE,
  append = TRUE,
  invname
)
}
\arguments{
\item{files}{character. Vector of file names with paths
(from the root directory).}

\item{root}{character. Root directory.}

\item{test}{logical. If TRUE, original investigation name will be extended
with the date-time to ensure unique name.}

\item{append}{logical. Should log entries be appended to the existing log?}

\item{invname}{character. Name that overrides automatic investigation name
if test is TRUE.}
}
\description{
Create an object of type 'documents' or 'data_files'
and upload file content as a blob.
}
\note{
Argument files is usually a result of a function skFilesToUpload
     that sets the option 'root' that was used.
}
\examples{
\dontrun{
skUploadFiles(files)
}

}
\seealso{
\code{\link{skFilesToUpload}}, \code{\link{skUpload}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
