% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{with_capture_groups_matrix}
\alias{with_capture_groups_matrix}
\title{Use capture groups in function-based replacements}
\usage{
with_capture_groups_matrix(fn)
}
\arguments{
\item{fn}{A function taking a single argument: a character matrix of capture groups,
where each row corresponds to a match, the first column is the full match and
subsequent columns are capture groups.}
}
\value{
A function identical to \code{fn}, but marked with an internal attribute
used by \code{\link[=compute_replacement]{compute_replacement()}} to dispatch on replacement logic.
}
\description{
This helper wraps a function to indicate that it should receive the full
capture group matrix as input, instead of the default vector of full matches.
This allows the replacement logic to use individual capture groups.
}
\details{
The capture matrix is the result of \code{\link[stringr:str_match_all]{stringr::str_match_all()}}: the first
column contains the full match, and subsequent columns contain capture groups.
}
\examples{
text <- "lorem ipsum foo_bar lorem ipsum bar_foo lorem ipsum"
fn_repl <- function(M) paste0(tolower(M[, 3L]), ".", toupper(M[, 2L]))
fn_repl <- with_capture_groups_matrix(fn_repl)
match_text(text, path = "example", pattern = "(\\\\w+)_(\\\\w+)", replacement = fn_repl)

}
