% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhCreate}
\alias{fhCreate}
\title{Create pISA layer or *fh* component.}
\usage{
fhCreate(type = "assays", meta)
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projets', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (Title, Description, *ToDo: add fields*).}
}
\value{
FAIRDOMhub created component.
}
\description{
Create pISA layer or *fh* component.
}
\note{
Upon success (status code 200) details
of newly created component can be used. Check status code.
}
\examples{
\\donotrun{
si <- fhCreate( type = "investigations"
  , meta= list(
      Title="Test Investigation"
    , Description="Some description")
    )
 str(si)
 sp <- fhCreate( type = "projects"
  , meta= list(
      Title="Test Project"
    , Description="Some description")
    )
 str(sp)
}
}
\seealso{
\code{\link{fhGET}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
