% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-api.R
\name{fhGet}
\alias{fhGet}
\title{Get inormation from repository.}
\usage{
fhGet(type, id, uri = options()$fhub$baseurl, ...)
}
\arguments{
\item{type}{Type of information (e.g. "person").}

\item{id}{Repository id of an item.}

\item{uri}{Repository base address (URI)..}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
Get inormation from repository.
}
\note{
Parameter ... is ignored at this time.
}
\examples{
\\donotrun{
fhIni()
options()$fhub$myid
r <- fhGet("people",options()$fhub$myid)
names(r)
r$response$status_code
status_code(r$response)
r
}
}
\seealso{
\code{\link{get ...}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
