% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{getRoot}
\alias{getRoot}
\title{Get root directory for pISA layer}
\usage{
getRoot(x = "p", path = getwd(), ...)
}
\arguments{
\item{x}{Character characteristic for pISA layer (one of p, I, S, or A).}

\item{path}{Path within the pISA-tree.}

\item{...}{Any other arguments.}
}
\value{
Relative path to the layer directory (from working directory).
}
\description{
Get root directory for pISA layer
}
\note{
The path should be compliant with the pISA-tree structure.
    Path defaults to the working directory, which is
    usually in or below an assay. Argument ... is not used.
}
\examples{
getRoot("p", path="d:/_p_prj/_I_inv/_S_st/_A_asy/other/doc")
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pISA}
