% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{print.pISAmeta}
\alias{print.pISAmeta}
\title{Print metadata object as Dlist}
\usage{
\method{print}{pISAmeta}(x, ...)
}
\arguments{
\item{x}{metadata object, data.frame with two columns}

\item{...}{any other arguments}
}
\description{
Print metadata object as Dlist
}
\note{
Metadata table is printed in convenient Dlist form.
}
\examples{
print(data.frame(Key=c("First","Second")
, Value=c("Description 1", "Description 2"))
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{package}
