% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{print.seek_api}
\alias{print.seek_api}
\title{Print method for seek_api object}
\usage{
\method{print}{seek_api}(x, content = FALSE)
}
\arguments{
\item{x}{Object of class \code{seek_api}.}

\item{content}{If FALSE (default), content is not printed.}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
Print method for seek_api object
}
\examples{
\donotrun{
skIni()
options()$fhub$myid
r <- skGet("people",options()$fhub$myid)
# Print contents
print( r, TRUE)
# Short version, default
r
r$response$status_code
status_code(r$response)
}

}
\seealso{
\code{\link{skParse ...}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
