% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisar-initialize.R
\name{readMeta}
\alias{readMeta}
\title{Read metadata file from the given directory}
\usage{
readMeta(x = ".", ...)
}
\arguments{
\item{x}{file path to the pISA layer}

\item{...}{any other arguments}
}
\value{
data frame with Key/value pairs with class 'pISAmeta'
}
\description{
Read metadata file from the given directory
}
\note{
Metadata table gets the class 'Dlist' to inherit a convenient print.
}
\examples{
\dontrun{
.pISAloc <- system.file("extdata","_p_Demo",package="pisar")
readMeta(.pISAloc)
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pISA}
