% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-package.R
\docType{package}
\name{seekr-package}
\alias{seekr}
\alias{seekr-package}
\title{seekr: Extract Matching Lines from Matching Files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a simple interface to recursively list files from a directory, filter them using a regular expression, read their contents, and extract lines that match a user-defined pattern. The package returns a tibble containing the matched lines, their line numbers, file paths, and the corresponding matched substrings. Designed for quick codebase exploration, log inspection, or any use case involving pattern-based file and line filtering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/smartiing/seekr}
  \item \url{https://smartiing.github.io/seekr/}
  \item Report bugs at \url{https://github.com/smartiing/seekr/issues}
}

}
\author{
\strong{Maintainer}: Sacha Martingay \email{martingay.sacha@hotmail.com}

}
\keyword{internal}
