% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skCreate}
\alias{skCreate}
\title{Create pISA layer or *sk* component.}
\usage{
skCreate(type = "assays", meta = list(), class = "EXP", file = "NA.TXT")
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projets', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (Title,
Description, *ToDo: add fields*).}

\item{class}{Assay class key string.
Possible values are 'EXP' and 'MODEL'.}

\item{file}{File name with path, relative to layer.}
}
\value{
FAIRDOMhub created component.
}
\description{
Create pISA layer or *sk* component.
}
\note{
Upon success (status code 200) details
of newly created component can be used. Check status code.

Created component id is set in options.
}
\examples{
\dontrun{
if(FALSE)
{
options(.sk$test)
options(sk.prid = 26)
skOptions("id")
meta= list(
      Title=paste("Test project", Sys.time())
    , Description="Testing upload")
meta
rm(sp)
 sp <- skCreate( type = "projects"
    , meta = meta
    )
    (r <- skFindId("projects",meta$Title))
 sp
 str(skData(sp))
# Call to API does not set the 'member' field
# Add member manually in the project page on the web site:
# /Actions/Administer project members
#
 options(sk.prid=26, sk.pid=sp$id)
 skOptions("id")
 meta= list(
      Title=paste("Test investigation", Sys.time())
    , Description="Testing upload")
    meta
    rm(si)
 si <- skCreate( type = "investigations"
  , meta=meta
    )
 si
 skData(si)$id
 skFindId("investigations",meta$Title)
 options(sk.iid=si$id)
 skOptions("id")
#'
 iid=skData(si)$id
 options(sk.prid=26, sk.pid=sp$id, sk.iid=si$id)
 skOptions("id")
 ss <- skCreate( type = "studies"
  , meta= list(
      Title=paste("Test study", Sys.time())
    , Description="Testing upload")
    )
 ss
 skData(ss)$id
#'
 options(sk.prid = 26
       , sk.pid=skData(sp)$id
       , sk.iid=skData(si)$id
       , sk.sid=skData(ss)$id
       )
 skOptions()
 sa <- skCreate( type = "assays"
  , meta= list(
      Title=paste("Test assay", Sys.time())
    , Description="Testing upload")
    , class="EXP"
    )
 str(sa)
 skData(sa)$relationships$submitter
 skData(sa)$links
}

# Type: data_file
astring <- "_p_Demo/_I_Test/_S_Show/_A_Work-R/"
oldwd <- setwd(system.file("extdata",astring,package="seekr"))
oldwd
.aname <- getLayer("A")
.aroot <- getRoot("A")
.ameta  <- readMeta()
 file <- "input/README.MD"
 options(sk.prid = 26, sk.pid=sp$id, sk.iid=si$id , sk.sid=ss$id , sk.aid=sa$id)
 skOptions()
 type <- "data_files"
 type <- "documents"
 sdat <- skCreate( type = type
  , meta= list(
      Title=paste("Test assay", Sys.time())
    , Description="Testing of upload")
  , file=file
    )
 #str(sdat)
 sdat
 skData(sdat)$id
if(interactive()) setwd(oldwd)
getwd()
res <- sdat$content
item_link <- file.path(res$meta$base_url,res$links$self)
if(interactive()) shell.exec(item_link)
}
}
\seealso{
\code{\link{skGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
