% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skExists}
\alias{skExists}
\title{Check if component exist.}
\usage{
skExists(type, title, verbose = FALSE)
}
\arguments{
\item{type}{Component name (possibly shortened e.g. 'pe' for 'people',
'proj' for 'projects', ...).}

\item{title}{Character string with the identifier
of the component (title part).}

\item{verbose}{logical, if TRUE, details are printed.}
}
\value{
logical value indicating whether the component named by
     argument title exists.
}
\description{
Check if component exist.
}
\examples{
\dontrun{
options(.sk$test)
skExists("people", "Andrej Blejec")
skExists("projects","Demo", verbose=TRUE)
skExists("pe", "Guest")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
