% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skExistsp}
\alias{skExistsp}
\title{Check if component exist within the project.}
\usage{
skExistsp(type, title, pid = options("sk.pid"), verbose = FALSE)
}
\arguments{
\item{type}{Component name (possibly shortened e.g. 'pe' for 'people',
'proj' for 'projects', ...).}

\item{title}{Character string with the identifier
of the component (title part).}

\item{verbose}{logical, if TRUE, details are printed.}
}
\value{
logical value indicating whether the component named by
     argument title exists.
}
\description{
Check if component exist within the project.
}
\examples{
\dontrun{
options(.sk$test)
skFindId("projects","_p_Demo")
skOptions()
skExistsp("people", "Andrej Blejec")
skExistsp("investigations","_I_Test", verbose=TRUE)
skExistsp("pe", "Guest")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
