% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skFilesToUpload}
\alias{skFilesToUpload}
\title{Prepare the list of files for upload}
\usage{
skFilesToUpload(
  root,
  exclude.nokey = FALSE,
  exclude.md = TRUE,
  verbose = TRUE,
  quiet = FALSE,
  skignore = file.path(root, "seekignore.txt"),
  recursive = TRUE
)
}
\arguments{
\item{root}{character string. Directory, pointing to the branch
of pISA-tree to be checked.}

\item{exclude.nokey}{logical value. If TRUE, the layer will be ignored
if the "Upload to FAIRDOMHub" is not present.}

\item{exclude.md}{logical value. If TRUE, the file will be ignored
if it contains just the header line.}

\item{verbose}{logical. Shold print progress details?}

\item{quiet}{logical. Should print final report?}

\item{skignore}{sharacter. File name or path to seekignore file.}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\description{
Prepare the list of files for upload
}
\note{
Function `skFilesToUpload` will first create a list
of files in root directory. Then, it will read the seekignore.txt
and extend it with patterns of layers that should not be uploaded
 according to metadata key "Upload to FAIRDOMHub".
In addition, seekignre list will be appended with the list of empty
README.MD files that might clutter the FAIRDOMHub list of files.
Then it will filter it according to declarations in the combined
seekignore pattern list.
}
\examples{
\dontrun{
skFilesToUpload(.sroot)
skFilesToUpload(.aroot)
dirs <- skFilesToUpload(.iroot)
dirs
skFilesToUpload(exclude.md=FALSE)
skFilesToUpload( verbose=FALSE, quiet=TRUE)
seekignore <- c("# ignore patterns","*.zip","*.txt")
skFilesToUpload(skignore=seekignore)
skFilesToUpload(file.path(.iroot),skignore=NULL)
}

}
\seealso{
\code{\link{skIgnoreBlankReadme}}, \code{\link{skDoNotUpload}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa,}
\keyword{seekr}
