% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skFindTitle}
\alias{skFindTitle}
\title{Get details of component with id from an *fh* object.}
\usage{
skFindTitle(type, id)
}
\arguments{
\item{type}{Components name (e.g. 'people', 'projets', ...).}

\item{id}{Character string with the identifier
of the component (id part).}
}
\value{
FAIRDOMhub component identifier: id, type and title.
    If argument title is missing, a data frame with identifiers
    for all items is returned. See note.
}
\description{
Get details of component with id from an *fh* object.
}
\note{
If item is not found, empty string is returned as title.
}
\examples{
\\donotrun{
skIni()
id <- skFindId("people","Guest")
id
idNew <- skFindTitle("people", id[1])
idNew
all.equal(id, idNew)
skFindTitle("people",0)
}
}
\seealso{
\code{\link{skFindTitle}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
