% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skIni}
\alias{skIni}
\title{Initialize FAIRDOMhub information}
\usage{
skIni(
  url = "https://www.fairdomhub.org/",
  myid = NULL,
  prid = NULL,
  pid = NULL,
  iid = NULL,
  sid = NULL,
  aid = NULL,
  test = TRUE
)
}
\arguments{
\item{test}{If TRUE, test server will be used..}
}
\value{
A list with URL and user information. For side effect see Notes.
}
\description{
Define FAIRDOMhub URL and user data
}
\note{
The returned list is added to the
     \code{options()} list under name 'seekr'.
}
\examples{
\dontrun{
skIni()
options("seekr")
#
skIni(test=FALSE)
options("seekr")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
