% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skIni}
\alias{skIni}
\title{Initialize FAIRDOMhub information}
\usage{
skIni(prid = NULL, pid = NULL, iid = NULL, sid = NULL, aid = NULL, test = TRUE)
}
\arguments{
\item{test}{If TRUE, test server will be used..}
}
\value{
A list with URL and user information. For side effect see Notes.
}
\description{
Define FAIRDOMhub URL and user data
}
\examples{
\dontrun{
skIni()
options("fhub")
#
skIni(test=FALSE)
options("fhub")
}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{file}
