% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skRead}
\alias{skRead}
\title{Read operation will return information about the instance identified.}
\usage{
skRead(type, id, uri = options("sk.url"), ...)
}
\arguments{
\item{type}{Type of information (e.g. "person").}

\item{id}{Repository id of an item.}

\item{uri}{Repository base address (URI).}

\item{...}{further arguments.}
}
\value{
An object (list) of class \code{seek_api}.
}
\description{
Read operation will return information about the instance identified.
}
\note{
Parameter ... is ignored at this time.

Touched component id is set in the options (see:
     \code\link{skOptions}).
}
\examples{
\dontrun{
options(.sk$test)
options("sk.myid")
r <- skRead("people",options("sk.myid"))
names(r)
r$response$status_code
status_code(r$response)
r
# Non existent user
skRead("people",0)
}
}
\seealso{
\code{\link{skRead}}, \code{\link{skGet}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
