% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekr-api.R
\name{skSkeleton}
\alias{skSkeleton}
\title{Create *sk* skeleton.}
\usage{
skSkeleton(type = "assay", meta, file)
}
\arguments{
\item{type}{Component name (e.g. 'people', 'projets', ...).}

\item{meta}{Data frame with pISA metadata or
a list with minimal information (Title, Description, *ToDo: add fields*).}
}
\value{
A list with the minimal information structure.
}
\description{
Creates *sk* object with required structure.
}
\seealso{
\code{\link{skCreate}}
}
\author{
Andrej Blejec \email{andrej.blejec@nib.si}
}
\keyword{pisa}
